requirements.txt
yfinance
mplfinance
pysqlite3
fastapi
uvicorn
pytest-cov